I've added an instance of a projectile to show how entity move works.
I've also a couple notes around it about how entities work. however I'll say it here too. Entities and Cells are both child classes of sf::FloatRect. sf::View returns a sf::FloatRect for the viewport. using this we can choose to only render what is on the players screen at any given time. by only rendering
the cells that intersect with the view port, and only rendering the entities that intersect with the view port. This should cut down on the time required to render each frame significantly later on, as when more entities are added there can be a lot happening on screen.