Added more to basic attack framework.
Basicly how I see it working is as follows, when a attack is issued a
projectile is fetched. The projectile takes an action and when on each
update the action knows what it's supposed to do. The projectile has a
time to live and a vector of where it should be going.

This vector may be updated by the action on update.

When the Time to live expires, the projectile should be ejected back
into the queue.

Actions may hold the base TTL or it could be in the projectile
constructor, I don't think it matters. HOWEVER, if it's in the
projectile that could make everything a lot simpler creation wise.